requirements.txt
numpy
pycryptodome
keras
tensorflow
termplot